fix: resolve conflicts with new arch
