+++ src/screens/novel/components/Info/NovelInfoHeader.tsx
@@ -87,6 +87,7 @@ const NovelInfoHeader = ({
   openDrawer,
   page,
   setCustomNovelCover,
+  saveNovelCover,
   theme,
   totalChapters,
   trackerSheetRef,
@@ -119,6 +120,7 @@ const NovelInfoHeader = ({
             setCustomNovelCover={
               isLoading ? showNotAvailable : setCustomNovelCover
             }
+            saveNovelCover={isLoading ? showNotAvailable : saveNovelCover}
           />
           <View style={styles.novelDetails}>
             <Row>
